# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ryuk",
    "version": "0.1.0"
  },
  "paths": {
    "/api/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Get Health",
        "operationId": "getHealth",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Health"
                }
              }
            }
          },
          "default": {
            "description": "Problem",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/models": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "Get Models",
        "operationId": "getModels",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RecognitionModelInfo"
                  },
                  "type": "array",
                  "title": "Response Getmodels"
                }
              }
            }
          },
          "default": {
            "description": "Problem",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/persons": {
      "get": {
        "tags": [
          "persons"
        ],
        "summary": "List Persons",
        "operationId": "listPersons",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "on_watchlist",
                "removed",
                "all"
              ],
              "type": "string",
              "default": "on_watchlist",
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PersonOfInterestSummary"
                  },
                  "title": "Response Listpersons"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "default": {
            "description": "Problem",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "persons"
        ],
        "summary": "Create Person",
        "operationId": "createPerson",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_createPerson"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonOfInterest"
                }
              }
            }
          },
          "409": {
            "description": "Warnings to acknowledge",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/WarningsProblem"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "default": {
            "description": "Problem",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/persons/{person_id}": {
      "get": {
        "tags": [
          "persons"
        ],
        "summary": "Get Person",
        "operationId": "getPerson",
        "parameters": [
          {
            "name": "person_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Person Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonOfInterest"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "default": {
            "description": "Problem",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "persons"
        ],
        "summary": "Update Person",
        "operationId": "updatePerson",
        "parameters": [
          {
            "name": "person_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Person Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonOfInterestChanges"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonOfInterest"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "default": {
            "description": "Problem",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/persons/{person_id}/photos": {
      "post": {
        "tags": [
          "persons"
        ],
        "summary": "Add Photo",
        "operationId": "addPhoto",
        "parameters": [
          {
            "name": "person_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Person Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_addPhoto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrolledPhoto"
                }
              }
            }
          },
          "409": {
            "description": "Warnings to acknowledge",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/WarningsProblem"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "default": {
            "description": "Problem",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/persons/{person_id}/photos/{photo_id}/image": {
      "get": {
        "tags": [
          "persons"
        ],
        "summary": "Get Photo Image",
        "operationId": "getPhotoImage",
        "parameters": [
          {
            "name": "person_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Person Id"
            }
          },
          {
            "name": "photo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Photo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The enrolled photo, upright and without metadata",
            "content": {
              "image/jpeg": {},
              "image/png": {},
              "image/webp": {}
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "default": {
            "description": "Problem",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/persons/{person_id}/photos/{photo_id}": {
      "delete": {
        "tags": [
          "persons"
        ],
        "summary": "Delete Photo",
        "operationId": "deletePhoto",
        "parameters": [
          {
            "name": "person_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Person Id"
            }
          },
          {
            "name": "photo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Photo Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "default": {
            "description": "Problem",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_addPhoto": {
        "properties": {
          "photo": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Photo"
          },
          "acknowledgedWarnings": {
            "items": {
              "$ref": "#/components/schemas/WarningCode"
            },
            "type": "array",
            "title": "Acknowledgedwarnings"
          }
        },
        "type": "object",
        "required": [
          "photo"
        ],
        "title": "Body_addPhoto"
      },
      "Body_createPerson": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "photo": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Photo"
          },
          "acknowledgedWarnings": {
            "items": {
              "$ref": "#/components/schemas/WarningCode"
            },
            "type": "array",
            "title": "Acknowledgedwarnings"
          }
        },
        "type": "object",
        "required": [
          "name",
          "photo"
        ],
        "title": "Body_createPerson"
      },
      "EnrolledPhoto": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "mediaType": {
            "type": "string",
            "title": "Mediatype"
          },
          "width": {
            "type": "integer",
            "title": "Width"
          },
          "height": {
            "type": "integer",
            "title": "Height"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "mediaType",
          "width",
          "height",
          "createdAt"
        ],
        "title": "EnrolledPhoto"
      },
      "Health": {
        "properties": {
          "status": {
            "type": "string",
            "const": "ok",
            "title": "Status"
          },
          "version": {
            "type": "string",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "status",
          "version"
        ],
        "title": "Health"
      },
      "ModelState": {
        "type": "string",
        "enum": [
          "active",
          "available",
          "not_evaluated",
          "unavailable"
        ]
      },
      "Network": {
        "type": "string",
        "enum": [
          "arcface",
          "facenet",
          "sface"
        ]
      },
      "PersonOfInterest": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "$ref": "#/components/schemas/PersonStatus"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "statusChangedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Statuschangedat"
          },
          "photos": {
            "items": {
              "$ref": "#/components/schemas/EnrolledPhoto"
            },
            "type": "array",
            "title": "Photos"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "status",
          "createdAt",
          "statusChangedAt",
          "photos"
        ],
        "title": "PersonOfInterest"
      },
      "PersonOfInterestChanges": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "PersonOfInterestChanges",
        "description": "What PATCH changes: only the name. Any other field, such as `status`, is refused with\n`422 invalid_request` rather than ignored."
      },
      "PersonOfInterestSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "$ref": "#/components/schemas/PersonStatus"
          },
          "photoCount": {
            "type": "integer",
            "title": "Photocount"
          },
          "coverPhotoId": {
            "type": "string",
            "title": "Coverphotoid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "statusChangedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Statuschangedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "status",
          "photoCount",
          "coverPhotoId",
          "createdAt",
          "statusChangedAt"
        ],
        "title": "PersonOfInterestSummary"
      },
      "PersonStatus": {
        "type": "string",
        "enum": [
          "on_watchlist",
          "removed"
        ]
      },
      "Provider": {
        "type": "string",
        "enum": [
          "cpu",
          "coreml"
        ]
      },
      "RecognitionModelInfo": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "network": {
            "$ref": "#/components/schemas/Network"
          },
          "provider": {
            "$ref": "#/components/schemas/Provider"
          },
          "weightsSha256": {
            "type": "string",
            "title": "Weightssha256"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "state": {
            "$ref": "#/components/schemas/ModelState"
          },
          "threshold": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Threshold"
          },
          "dimension": {
            "type": "integer",
            "title": "Dimension"
          },
          "msPerFace": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Msperface"
          }
        },
        "type": "object",
        "required": [
          "id",
          "network",
          "provider",
          "weightsSha256",
          "name",
          "state",
          "threshold",
          "dimension",
          "msPerFace"
        ],
        "title": "RecognitionModelInfo"
      },
      "WarningCode": {
        "enum": [
          "duplicate_name",
          "looks_like_other",
          "may_not_be_same_person"
        ],
        "type": "string"
      },
      "EnrollmentWarning": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/WarningCode"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "personId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Personid"
          }
        },
        "required": [
          "code",
          "detail",
          "personId"
        ],
        "title": "EnrollmentWarning",
        "type": "object"
      },
      "Problem": {
        "description": "An RFC 9457 problem detail with Ryuk's `code` extension.",
        "properties": {
          "type": {
            "title": "Type",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "code": {
            "title": "Code",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code"
        ],
        "title": "Problem",
        "type": "object"
      },
      "WarningsProblem": {
        "description": "`409 warnings`: resend with every listed code in `acknowledgedWarnings` to proceed.",
        "properties": {
          "type": {
            "title": "Type",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/EnrollmentWarning"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code",
          "warnings"
        ],
        "title": "WarningsProblem",
        "type": "object"
      }
    }
  }
}